+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 ## 2023
+* `feb-mar`
+	* added new `sentry` advanced image
 * `jan`
 	* added new `jira` releases
 	* added new `bitbucket` releases
 	* added new `confluence` releases
+	* added *first* `crowd` release (version `5.1.1`)
 	* fixed `git` binary sources
 -------------------------------------------------------------------
 ## 2022